refactor(core): merge dotenv helper into env helper
